Update user dashboard
