+++ src/components/css/styles.css
@@ -76,9 +76,6 @@ a {
 #dropnav a {
     color: #007bff;}
 
-#no_loan {
-    margin:5%
-}
 .user {
     color: maroon
 }
